Exception handling

+++ src/main/java/com/vilaka/ecommerce/services/ProductServices.java
@@ -3,14 +3,16 @@ package com.vilaka.ecommerce.services;
 import com.vilaka.ecommerce.dto.ProductDTO;
 import com.vilaka.ecommerce.entities.Product;
 import com.vilaka.ecommerce.repositories.ProductRepository;
+import com.vilaka.ecommerce.services.exceptions.DatabaseException;
 import com.vilaka.ecommerce.services.exceptions.ResourceNotFoundException;
+import jakarta.persistence.EntityNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
-
-import java.util.Optional;
 
 @Service
 public class ProductServices {
@@ -18,7 +20,7 @@ public class ProductServices {
     private ProductRepository repository;
 
     @Transactional(readOnly = true)
-    public ProductDTO findById(Long id){
+    public ProductDTO findById(Long id) {
         Product product = repository.findById(id).orElseThrow(
                 () -> new ResourceNotFoundException("Recurso não encontrado!"));
         return new ProductDTO(product);
@@ -35,13 +37,13 @@ public class ProductServices {
      */
 
     @Transactional(readOnly = true)
-    public Page<ProductDTO> findAll(Pageable pageable){
+    public Page<ProductDTO> findAll(Pageable pageable) {
         Page<Product> result = repository.findAll(pageable);
         return result.map(x -> new ProductDTO(x));
     }
 
     @Transactional
-    public ProductDTO insert(ProductDTO dto){
+    public ProductDTO insert(ProductDTO dto) {
         Product entity = new Product();
         copyDtoToEntity(dto, entity);
         entity = repository.save(entity);
@@ -49,18 +51,32 @@ public class ProductServices {
     }
 
     @Transactional
-    public ProductDTO update(Long id, ProductDTO dto){
+    public ProductDTO update(Long id, ProductDTO dto) {
+        try {
+            Product entity = repository.getReferenceById(id);
+            copyDtoToEntity(dto, entity);
+            entity = repository.save(entity);
+            return new ProductDTO(entity);
+        } catch (EntityNotFoundException e) {
+            throw new ResourceNotFoundException("Este id não existe!");
+        }
 
-        Product entity = repository.getReferenceById(id);
-        copyDtoToEntity(dto, entity);
-        entity = repository.save(entity);
-        return new ProductDTO(entity);
 
     }
 
-    @Transactional
-    public void delete(Long id){
-        repository.deleteById(id);
+    @Transactional(propagation = Propagation.SUPPORTS)
+    public void delete(Long id) {
+        if (repository.existsById(id)) {
+            try{
+                repository.deleteById(id);
+            }
+            catch (DataIntegrityViolationException e){
+                throw new DatabaseException("Falha de integridade referencial!");
+            }
+        }
+        else {
+            throw new ResourceNotFoundException("Recurso não encontrado para deletar!");
+        }
     }
 
     private void copyDtoToEntity(ProductDTO dto, Product entity) {
